+++ packages/browser-wallet/src/popup/pages/ViewSeedPhrase/i18n/en.ts
@@ -1,0 +1,9 @@
+const t = {
+    title: 'View secret recovery phrase',
+    description: 'Please enter your passcode to show the secret recovery phrase.',
+    copyDescription: 'Press the button to copy your secret recovery phrase.',
+    show: 'Show secret recovery phrase',
+    done: 'Done',
+};
+
+export default t;