+++ packages/browser-wallet/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.3.2
 
+### Added
+
+-   A page in settings to display the secret recovery phrase.
+
 ### Fixed
 
 -   Updating stake crashing if the amount as microCCD exceeds 64 bits.